+++ processor/src/test/fixtures/expected/com/example/deprecation/Arez_DeprecationModel.java
@@ -1,0 +1,348 @@
+package com.example.deprecation;
+
+import arez.ActionFlags;
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.ComputableValue;
+import arez.Disposable;
+import arez.Locator;
+import arez.ObservableValue;
+import arez.Observer;
+import arez.SafeProcedure;
+import arez.component.DisposeNotifier;
+import arez.component.Identifiable;
+import arez.component.Verifiable;
+import arez.component.internal.ComponentKernel;
+import arez.component.internal.MemoizeCache;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.processing.Generated;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+@SuppressWarnings("unchecked")
+final class Arez_DeprecationModel extends DeprecationModel implements Disposable, Identifiable<Integer>, Verifiable, DisposeNotifier {
+  private static volatile int $$arezi$$_nextId;
+
+  private final ComponentKernel $$arezi$$_kernel;
+
+  @Nonnull
+  private final ObservableValue<Long> $$arez$$_value;
+
+  @Nonnull
+  private final MemoizeCache<Long> $$arez$$_count;
+
+  @Nonnull
+  private final MemoizeCache<Long> $$arez$$_time2;
+
+  @Nonnull
+  private final ComputableValue<Long> $$arez$$_time;
+
+  @Nonnull
+  private final Observer $$arez$$_myObserve;
+
+  @Nonnull
+  private final Observer $$arez$$_render;
+
+  @Nonnull
+  private final Observer $$arez$$_render2;
+
+  @Nullable
+  private DeprecationModel.MyEntity $$arezr$$_Blah;
+
+  Arez_DeprecationModel() {
+    super();
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( Arez::areReferencesEnabled, () -> "Attempted to create instance of component of type 'com_example_deprecation_DeprecationModel' that contains references but Arez.areReferencesEnabled() returns false. References need to be enabled to use this component" );
+    }
+    final ArezContext $$arezv$$_context = Arez.context();
+    final int $$arezv$$_id = ++$$arezi$$_nextId;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "com_example_deprecation_DeprecationModel." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "com_example_deprecation_DeprecationModel", $$arezv$$_id, $$arezv$$_name, this::$$arezi$$_nativeComponentPreDispose ) : null;
+    this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_dispose, null, true, false, false );
+    this.$$arez$$_value = $$arezv$$_context.observable( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".value" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> ( this.$$arezi$$_kernel.isNotReady() ? null : super.getValue() ) : null, Arez.arePropertyIntrospectorsEnabled() ? v -> super.setValue( v ) : null );
+    this.$$arez$$_count = new MemoizeCache<>( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".count" : null, args -> super.count((long) args[ 0 ], (float) args[ 1 ]), 2, ComputableValue.Flags.AREZ_DEPENDENCIES );
+    this.$$arez$$_time2 = new MemoizeCache<>( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".time2" : null, args -> $$arezi$$_memoize_time2((String) args[ 0 ]), 1, ComputableValue.Flags.AREZ_DEPENDENCIES );
+    this.$$arez$$_time = $$arezv$$_context.computable( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".time" : null, () -> super.getTime(), this::onTimeActivate, this::onTimeDeactivate, ComputableValue.Flags.AREZ_DEPENDENCIES | ComputableValue.Flags.RUN_LATER );
+    this.$$arez$$_myObserve = $$arezv$$_context.observer( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".myObserve" : null, () -> super.myObserve(), Observer.Flags.RUN_LATER | Observer.Flags.NESTED_ACTIONS_DISALLOWED | Observer.Flags.AREZ_DEPENDENCIES );
+    this.$$arez$$_render = $$arezv$$_context.tracker( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".render" : null, () -> super.onRenderDepsChange(), Observer.Flags.RUN_LATER | Observer.Flags.NESTED_ACTIONS_DISALLOWED | Observer.Flags.AREZ_DEPENDENCIES );
+    this.$$arez$$_render2 = $$arezv$$_context.observer( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".render2" : null, () -> super.render2(), () -> super.onRender2DepsChange(), Observer.Flags.RUN_LATER | Observer.Flags.NESTED_ACTIONS_DISALLOWED | Observer.Flags.AREZ_DEPENDENCIES );
+    this.$$arezi$$_kernel.componentConstructed();
+    this.$$arezi$$_link_Blah();
+    this.$$arezi$$_kernel.componentComplete();
+  }
+
+  @Nonnull
+  private Locator $$arezi$$_locator() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named '$$arezi$$_locator' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    return this.$$arezi$$_kernel.getContext().locator();
+  }
+
+  private int $$arezi$$_id() {
+    return this.$$arezi$$_kernel.getId();
+  }
+
+  @Override
+  @Nonnull
+  public Integer getArezId() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'getArezId' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'getArezId' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    return $$arezi$$_id();
+  }
+
+  private void $$arezi$$_nativeComponentPreDispose() {
+    this.$$arezi$$_kernel.notifyOnDisposeListeners();
+  }
+
+  @Override
+  public void addOnDisposeListener(@Nonnull final Object key, @Nonnull final SafeProcedure action) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'addOnDisposeListener' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    this.$$arezi$$_kernel.addOnDisposeListener( key, action );
+  }
+
+  @Override
+  public void removeOnDisposeListener(@Nonnull final Object key) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'removeOnDisposeListener' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'removeOnDisposeListener' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    this.$$arezi$$_kernel.removeOnDisposeListener( key );
+  }
+
+  @Override
+  public boolean isDisposed() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'isDisposed' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'isDisposed' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    return this.$$arezi$$_kernel.isDisposed();
+  }
+
+  @Override
+  public void dispose() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'dispose' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'dispose' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    this.$$arezi$$_kernel.dispose();
+  }
+
+  private void $$arezi$$_dispose() {
+    this.$$arez$$_myObserve.dispose();
+    this.$$arez$$_render.dispose();
+    this.$$arez$$_render2.dispose();
+    this.$$arez$$_count.dispose();
+    this.$$arez$$_time2.dispose();
+    this.$$arez$$_time.dispose();
+    this.$$arez$$_value.dispose();
+  }
+
+  @Override
+  public void verify() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'verify' invoked on uninitialized component of type 'com_example_deprecation_DeprecationModel'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'verify' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'verify' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() && Arez.isVerifyEnabled() ) {
+      Guards.apiInvariant( () -> this == $$arezi$$_locator().findById( DeprecationModel.class, $$arezi$$_id() ), () -> "Attempted to lookup self in Locator with type DeprecationModel and id '" + $$arezi$$_id() + "' but unable to locate self. Actual value: " + $$arezi$$_locator().findById( DeprecationModel.class, $$arezi$$_id() ) );
+      final int $$arezv$$_BlahId = this.getMyEntityId();
+      final DeprecationModel.MyEntity $$arezv$$_Blah = this.$$arezi$$_locator().findById( DeprecationModel.MyEntity.class, $$arezv$$_BlahId );
+      Guards.apiInvariant( () -> null != $$arezv$$_Blah, () -> "Reference named 'Blah' on component named '" + this.$$arezi$$_kernel.getName() + "' is unable to resolve entity of type com.example.deprecation.DeprecationModel.MyEntity and id = " + getMyEntityId() );
+    }
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  public long getValue() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getValue' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    this.$$arez$$_value.reportObserved();
+    return super.getValue();
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  void setValue(final long time) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'setValue' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    this.$$arez$$_value.preReportChanged();
+    final long $$arezv$$_currentValue = super.getValue();
+    if ( time != $$arezv$$_currentValue ) {
+      super.setValue( time );
+      this.$$arez$$_value.reportChanged();
+    }
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  void myObserve() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.fail( () -> "Observe method named 'myObserve' invoked but @Observe(executor=INTERNAL) annotated methods should only be invoked by the runtime." );
+    }
+    super.myObserve();
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  void render() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'render' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    this.$$arezi$$_kernel.getContext().safeObserve( this.$$arez$$_render, () -> super.render(), null );
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  void render2() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.fail( () -> "Observe method named 'render2' invoked but @Observe(executor=INTERNAL) annotated methods should only be invoked by the runtime." );
+    }
+    super.render2();
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  @Nonnull
+  Observer getRender2Observer() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getRender2Observer' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    return $$arez$$_render2;
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  public void doStuff(final long time, final float someOtherParameter) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'doStuff' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    this.$$arezi$$_kernel.getContext().safeAction( Arez.areNamesEnabled() ? this.$$arezi$$_kernel.getName() + ".doStuff" : null, () -> super.doStuff( time, someOtherParameter ), ActionFlags.READ_WRITE | ActionFlags.VERIFY_ACTION_REQUIRED, Arez.areSpiesEnabled() ? new Object[] { time, someOtherParameter } : null );
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  public long count(final long time, final float someOtherParameter) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'count' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    return this.$$arez$$_count.get( time, someOtherParameter );
+  }
+
+  public long $$arezi$$_memoize_time2(final String $$arezi$$_myContextVar) {
+    try {
+      pushMyContextVar( $$arezi$$_myContextVar );
+      return super.getTime2();
+    } finally {
+      popMyContextVar( $$arezi$$_myContextVar );
+    }
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  public long getTime2() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getTime2' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    return this.$$arez$$_time2.get( captureMyContextVar() );
+  }
+
+  @Override
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  public long getTime() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getTime' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    return this.$$arez$$_time.get();
+  }
+
+  @SuppressWarnings({
+      "RedundantSuppression",
+      "deprecation"
+  })
+  @Override
+  DeprecationModel.MyEntity getMyEntity() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getMyEntity' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != $$arezr$$_Blah, () -> "Nonnull reference method named 'getMyEntity' invoked on component named '" + this.$$arezi$$_kernel.getName() + "' but reference has not been resolved yet is not lazy. Id = " + getMyEntityId() );
+    }
+    return this.$$arezr$$_Blah;
+  }
+
+  private void $$arezi$$_link_Blah() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named '$$arezi$$_link_Blah' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + this.$$arezi$$_kernel.getName() + "'" );
+    }
+    final int id = this.getMyEntityId();
+    this.$$arezr$$_Blah = this.$$arezi$$_locator().findById( DeprecationModel.MyEntity.class, id );
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != $$arezr$$_Blah, () -> "Reference named 'Blah' on component named '" + this.$$arezi$$_kernel.getName() + "' is unable to resolve entity of type com.example.deprecation.DeprecationModel.MyEntity and id = " + getMyEntityId() );
+    }
+  }
+
+  private void $$arezi$$_delink_Blah() {
+    this.$$arezr$$_Blah = null;
+  }
+
+  @Override
+  public String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + this.$$arezi$$_kernel.getName() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}